Hable complex authentication

+++ src/Controller/PostController.php
@@ -15,45 +15,49 @@ class PostController extends AbstractController
     #[Route('/', name: 'home')]
     public function index(ManagerRegistry $doctrine ): Response
     {
-        //on récupère d'abord le Pository
-        $repository = $doctrine->getRepository(Post::class);
-        $posts = $repository->findAll();// SELECT * FROM 'post'
-        return $this->render('post/index.html.twig', [
+            //on récupère d'abord le Repository
+            $repository = $doctrine->getRepository(Post::class);
+            $posts = $repository->findAll();// SELECT * FROM 'post'
+            return $this->render('post/index.html.twig', [
             //on envoie au template tous les posts
             "posts" => $posts 
-        ]);
+            ]);
     }
 
     #[Route('/post/new')]
     // on injecte la requête HTTP
     public function create(Request $request, ManagerRegistry $doctrine): Response
     {
-        // création du formulaire
-        $post = new Post();
-        $form = $this->createForm(PostType::class, $post);
+        if ($this->getUser()) {
+            // création du formulaire
+            $post = new Post();
+            $form = $this->createForm(PostType::class, $post);
 
-        // mise à jour de l'objet $form avec les valeurs saisies
-        $form->handleRequest($request);
-        
-        //on s'assure de la validité du formulaire et que les valeurs sont cohérentes
-        if ($form->isSubmitted() && $form->isValid()) {
-            //on associe le post avec l'id de l'user qui l'a écrit
-            $post->setUser($this->getUser());
+            // mise à jour de l'objet $form avec les valeurs saisies
+            $form->handleRequest($request);
             
-            $em = $doctrine->getManager();
-            $em->persist($post);
-            $em->flush();
-            // on redirige vers la page d'accueil
-            return $this->redirectToRoute('home');
+            //on s'assure de la validité du formulaire et que les valeurs sont cohérentes
+            if ($form->isSubmitted() && $form->isValid()) {
+                //on associe le post avec l'id de l'user qui l'a écrit
+                $post->setUser($this->getUser());
+                $em = $doctrine->getManager();
+                $em->persist($post);
+                $em->flush();
+                // on redirige vers la page d'accueil
+                return $this->redirectToRoute('home');
+            }
+            return $this->render('user/form.html.twig', [
+                "form" => $form->createView()
+            ]);
+        } else {
+            return $this->redirectToRoute("login");
         }
-        return $this->render('user/form.html.twig', [
-            "form" => $form->createView()
-        ]);
     }
 
     #[Route('/post/delete/{id<\d+>}', name:'delete_post')]
     public function delete(Post $post, ManagerRegistry $doctrine): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $em = $doctrine->getManager();
         $em->remove($post);
         $em->flush();
@@ -63,6 +67,7 @@ class PostController extends AbstractController
     #[Route('/post/edit/{id<\d+>}', name:'edit_post')]
     public function update(Post $post, Request $request, ManagerRegistry $doctrine): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $form = $this->createForm(PostType::class, $post);
         $form->handleRequest($request);
         
@@ -72,7 +77,7 @@ class PostController extends AbstractController
             return $this->redirectToRoute('home');
         }
         return $this->render('post/form.html.twig', [
-            "post_form" => $form->createView()
+            "form" => $form->createView()
         ]);
     }
 
@@ -80,6 +85,7 @@ class PostController extends AbstractController
     // on injecte la requête HTTP
     public function copy(Post $post,ManagerRegistry $doctrine): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $copyPost = clone $post;
 
         $em = $doctrine->getManager();
